MAGE-1104 Relocate legacy Logger class to separate namespace
