+++ Helper/Image.php
@@ -2,6 +2,7 @@
 
 namespace Algolia\AlgoliaSearch\Helper;
 
+use Algolia\AlgoliaSearch\Logger\Logger;
 use Magento\Catalog\Model\Product\ImageFactory;
 use Magento\ConfigurableProduct\Model\Product\Type\Configurable as ProductTypeConfigurable;
 use Magento\Framework\App\Helper\Context;
